+++ duva/src/domains/cluster_actors/consensus/election.rs
@@ -1,4 +1,4 @@
-use crate::domains::{cluster_actors::replication::Role, peers::identifier::PeerIdentifier};
+use crate::domains::{cluster_actors::replication::ReplicationRole, peers::identifier::PeerIdentifier};
 
 #[derive(Debug, Clone)]
 pub(crate) enum ElectionState {
@@ -8,10 +8,10 @@ pub(crate) enum ElectionState {
 }
 
 impl ElectionState {
-    pub(crate) fn new(role: &Role) -> Self {
+    pub(crate) fn new(role: &ReplicationRole) -> Self {
         match role {
-            Role::Leader => ElectionState::Leader,
-            Role::Follower => ElectionState::Follower { voted_for: None },
+            ReplicationRole::Leader => ElectionState::Leader,
+            ReplicationRole::Follower => ElectionState::Follower { voted_for: None },
         }
     }
     pub(crate) fn become_leader(&mut self) {
