+++ duva/src/domains/peers/mod.rs
@@ -2,4 +2,5 @@ pub mod connected_peer_info;
 pub(crate) mod connected_types;
 pub mod identifier;
 
+pub mod cluster_peer;
 pub(crate) mod peer;
