+++ duva/src/lib.rs
@@ -13,7 +13,7 @@ use domains::caches::cache_manager::CacheManager;
 use domains::cluster_actors::ClusterActor;
 use domains::cluster_actors::commands::ClusterCommand;
 use domains::cluster_actors::replication::ReplicationState;
-use domains::cluster_actors::replication::Role;
+use domains::cluster_actors::replication::ReplicationRole;
 use domains::config_actors::config_manager::ConfigManager;
 use domains::saves::snapshot::snapshot_loader::SnapshotLoader;
 pub use init::Environment;
@@ -129,7 +129,7 @@ impl StartUpFacade {
             peers.push(PeerIdentifier(self.registry.config_manager.bind_addr()));
 
             let is_leader =
-                self.registry.cluster_communication_manager().role().await? == Role::Leader;
+                self.registry.cluster_communication_manager().role().await? == ReplicationRole::Leader;
             let Ok((reader, writer)) = authenticate(stream, peers, is_leader).await else {
                 eprintln!("[ERROR] Failed to authenticate client stream");
                 continue;
